requirements.txt
# FastAPI & Web
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings

# Database
SQLAlchemy
psycopg2-binary
alembic
python-dotenv

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]

# Logging
loguru

# ML & Data (CPU-only versions)
numpy
pandas
faker
scikit-learn==1.3.2
xgboost
lightgbm
joblib
matplotlib
seaborn

# Advanced ML Libraries
# Time Series Forecasting
prophet>=1.1.4
statsmodels>=0.14.0
arch>=6.2.0  # For GARCH models
pmdarima>=2.0.4  # Auto-ARIMA

# Deep Learning (CPU-only)
torch>=2.0.0 --index-url https://download.pytorch.org/whl/cpu
torchvision>=0.15.0 --index-url https://download.pytorch.org/whl/cpu
tensorflow-cpu>=2.15.0
keras>=3.0.0

# Advanced ML Models
catboost>=1.2.2
imbalanced-learn>=0.11.0
surprise>=1.1.3  # Collaborative filtering
implicit>=0.7.2  # Matrix factorization

# Feature Engineering & Selection
feature-engine>=1.6.2
boruta>=0.3
shap>=0.43.0
lime>=0.2.0.1

# Time Series & External Data
yfinance>=0.2.22  # Market data
holidays>=0.34  # Holiday data
geopy>=2.4.0  # Geolocation
weatherapi>=1.0.0  # Weather data (if available)

# Streaming & Real-time
kafka-python>=2.0.2
celery>=5.3.4
redis-py-cluster>=2.1.3

# Experiment tracking & model management
mlflow-skinny==2.16.1
wandb>=0.16.0

# Additional ML dependencies for tracking (CPU-only)
tensorboard
optuna>=3.4.0  # For hyperparameter optimization
hyperopt>=0.2.7
bayesian-optimization>=1.4.3

# Caching and Performance
redis
hiredis

# HTTP Client
httpx
requests
aiofiles

# Development and Testing
pytest
pytest-asyncio
pytest-cov

# Utilities
typing-extensions

# Explicitly avoid CUDA packages
--extra-index-url https://download.pytorch.org/whl/cpu